Extract shared benchmark utilities and remove dead scripts
- Create scripts/performance/benchmark_base.py with shared data loading,
  enrichment execution, baseline I/O, and JSON output helpers
- Refactor benchmark_enrichment.py and detect_performance_regression.py to
  use benchmark_base, eliminating ~760 lines of duplicated code
- Remove scripts/data/extract_recipient_lookup.py (deprecated, replaced by
  stream_extract_recipient.py)
- Remove scripts/data/parallel_download.py (dead code, unused by any script)
